--- src/main/java/org/devopology/test/engine/TestEngineException.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2022-2023 Douglas Hoard
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.devopology.test.engine;
-
-public class TestEngineException extends RuntimeException {
-
-    public TestEngineException(String message) {
-        super(message);
-    }
-
-    public TestEngineException(String message, Throwable throwable) {
-        super(message, throwable);
-    }
-}
