--- src/main/java/org/devopology/test/engine/internal/descriptor/TestEngineTestSource.java
@@ -1,17 +1,0 @@
-package org.devopology.test.engine.internal.descriptor;
-
-import org.junit.platform.engine.TestDescriptor;
-import org.junit.platform.engine.TestSource;
-
-public class TestEngineTestSource implements TestSource {
-
-    private TestDescriptor testDescriptor;
-
-    public TestEngineTestSource(TestDescriptor testDescriptor) {
-        this.testDescriptor = testDescriptor;
-    }
-
-    public TestDescriptor getTestDescriptor() {
-        return testDescriptor;
-    }
-}
